Precision functions to utils
