+++ packages/office-ui-fabric-react/src/components/SpinButton/SpinButton.test.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import * as ReactTestUtils from 'react-dom/test-utils';
 import * as renderer from 'react-test-renderer';
-import { SpinButton, precisionRound, calculatePrecision } from './SpinButton';
+import { SpinButton } from './SpinButton';
 import { KeyCodes } from '../../Utilities';
 
 describe('SpinButton', () => {
@@ -643,43 +643,5 @@ describe('SpinButton', () => {
     );
 
     expect(onIncrement).toBeCalled();
-  });
-
-  it('caluclatePrecision should work as intended', () => {
-    expect(calculatePrecision(0)).toEqual(0);
-    expect(calculatePrecision(1)).toEqual(0);
-    expect(calculatePrecision('1')).toEqual(0);
-
-    expect(calculatePrecision(200)).toEqual(-2);
-    expect(calculatePrecision(32100012300000)).toEqual(-5);
-
-    expect(calculatePrecision(231.00)).toEqual(0);
-    expect(calculatePrecision('231.00')).toEqual(2);
-    expect(calculatePrecision(321.00002)).toEqual(5);
-    expect(calculatePrecision('321.00002')).toEqual(5);
-
-    expect(calculatePrecision(.002)).toEqual(3);
-    expect(calculatePrecision('.002')).toEqual(3);
-  });
-
-  it('precisionRound should work as intended', () => {
-    expect(precisionRound(1234, 0)).toEqual(1234);
-    expect(precisionRound(1234, -1)).toEqual(1230);
-    expect(precisionRound(1234, -3)).toEqual(1000);
-
-    expect(precisionRound(1234.5678, 0)).toEqual(1235);
-    expect(precisionRound(1234.5678, 1)).toEqual(1234.6);
-    expect(precisionRound(1234.5678, 5)).toEqual(1234.5678);
-
-    expect(precisionRound(1234.555, 2)).toEqual(1234.56);
-    expect(precisionRound(1234.554, 2)).toEqual(1234.55);
-    expect(precisionRound(1250, -2)).toEqual(1300);
-    expect(precisionRound(1249, -2)).toEqual(1200);
-
-    // Different bases
-    expect(precisionRound(1234.5, -2, 2)).toEqual(1236);
-    expect(precisionRound(1234.5, -2, 16)).toEqual(1280);
-    expect(precisionRound(1234.5, -2, 8)).toEqual(1216);
-    expect(precisionRound(1234.5, -2, 7)).toEqual(1225);
   });
 });
